Refactor: extract `TaskFinalize` as standalone component
