+++ src/modules/dashboard/components/TaskFinalize/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TaskFinalize';
